+++ Changelog.md
@@ -1,6 +1,7 @@
 ### 0.5.14 (unreleased)
 
 Language Features:
+ * Allow to obtain the selector of public or external library functions via a member ``.selector``.
 
 
 Compiler Features: